+++ src/cli/mod.rs
@@ -5,10 +5,13 @@ macro_rules! build_env_name {
     }};
 }
 pub(crate) use build_env_name;
+
 pub mod app_args;
 
 pub mod deletion_arguments;
+pub mod edit_argument;
 pub mod edit_command;
+pub mod edit_year;
 pub mod list_command;
 pub mod month_edit_command;
 pub mod month_list_command;