+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] 
 
+## [0.5.3] - 2023.02.04
+
 ### Added 
 
 - Editor for opening journals can adjusted CLI option, environment variable and option in configuration file
@@ -28,7 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.5.0] - 2023.02.01
 
-[unreleased]: https://github.com/BoolPurist/daily_ruster_man/compare/v0.5.1...HEAD
+[unreleased]: https://github.com/BoolPurist/daily_ruster_man/compare/v0.5.3...HEAD
+[0.5.3]: https://github.com/BoolPurist/daily_ruster_man/compare/v0.5.1...v0.5.3
 [0.5.1]: https://github.com/BoolPurist/daily_ruster_man/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/BoolPurist/daily_ruster_man/releases/tag/v0.5.0